% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{array_to_table}
\alias{array_to_table}
\title{Converts an array into a "long" format data frame}
\usage{
array_to_table(arr, index_colnames = default_index_colnames(arr))
}
\arguments{
\item{array}{An R array, with no more than 15 dimensions}
}
\value{
A \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
This function is designed to allow you to convert an array first into a data
frame, allowing you to then save it using a function such as \code{\link[arrow:write_parquet]{arrow::write_parquet()}} or
\code{\link[utils:write.table]{utils::write.csv()}}. Then, you can open the resulting file in a compatible
database (notably DuckDB), which you can use to construct an \code{\link[=SqlArray]{SqlArray()}}.
}
\examples{
rnorm(n = 16) |>
 matrix(ncol = 4) |>
 delaytional::array_to_table()
}
