% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{calculate_dims}
\alias{calculate_dims}
\title{Returns a named integer vector that describes the length of each dimension}
\usage{
calculate_dims(index_cols, tbl)
}
\arguments{
\item{index_cols}{A character vector of the columns in \code{tbl} that indicate
the indices within the array. For example \code{index_cols=c("a", "b")} means
that \code{tbl} has at least the columns \code{a} and \code{b}, where \code{a} contains the row
indices of each array entry, and \code{b} contains the column indices of the
entries. You will need to use a longer vector to support higher dimensional
arrays. Note that all of these columns \emph{must} be integer type columns.}

\item{tbl}{An SQL-backed data frame, probably created using
\code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl.src_dbi()}}. The table that you use must store each "cell" of
the array in a different row of the table, with different columns indicating
the indices and value of that cell. 0 values do not need to be represented
in the table (ie sparse representation is supported), although they can be
present if you want.}
}
\description{
Returns a named integer vector that describes the length of each dimension
}
