% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{default_index_colnames}
\alias{default_index_colnames}
\title{Generates some default column names for the index columns}
\usage{
default_index_colnames(arr)
}
\arguments{
\item{arr}{Some native R array}
}
\description{
Generates some default column names for the index columns
}
\keyword{internal}
