% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{store_sql_array}
\alias{store_sql_array}
\title{Saves an array-like to the database as a table}
\usage{
store_sql_array(arr, index_colnames = default_index_colnames(arr), ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[dplyr:copy_to]{dplyr::copy_to}}
  \describe{
    \item{\code{dest}}{remote data source}
    \item{\code{name}}{name for new remote table.}
    \item{\code{overwrite}}{If \code{TRUE}, will overwrite an existing table with
name \code{name}. If \code{FALSE}, will throw an error if \code{name} already
exists.}
  }}
}
\value{
An \linkS4class{SqlArray} instance
}
\description{
Saves an array-like to the database as a table
}
